[cirrus] Implement BoostFunctionFeature
Use one class per feature type. While not strictly needed it helps
clarify what are the different responsibilities. We had 4 keywords in
the same class that had to deal with different concepts:
 - 2 boosts
 - 2 filters
 - 2 that require db access

Since the new parser now requires that individual steps are properly
declared I think it helps to reduce branching in the various methods
like parseValue/expand/getBoostFunctionBuilder and soon getFilter.

This also drops support negating boosting function, I don't think it
worked well. It was applying a boost of 0.1 when negated but I think
we should do the reciprocal to the distance.

Bug: T195305
Depends-On: I6b0cb60db2440ea4ab4b4f0c9247385b77fa98df
Change-Id: I4babe7eab706f974c4195f28be39141332b05819