Use the new LinksUpdate::getPageId()
I want to deprecate all public property access in LinksUpdate.

I removed the guard for zero mId since it seems redundant with the
identical code in LinksUpdate::doUpdate(). Maybe that code was
duplicated from core in order to support previous versions of core. But
this change depends on getPageId() and hence depends on a version of
core with the guard.

Bug: T296895
Change-Id: I849b8b8055966971612898078dd7b28081ab7f67