Convert to abstract schema
This adds the postgres schema to the extension

This unify both schema files.
The difference is about the columns gt_lat_int/gt_lon_int only used for
config GeoDataBackend = db.
To simplify handling for configs the difference is removed, that means
for GeoDataBackend = elasticsearch the fields are unused and only stores
null.
For old installs both fields are added by the updater.

Bug: T268550
Depends-On: If55fdf89e1050a0bada4f963e600b5e73ffc01bf
Change-Id: If022d2cd9f4ded58798de825d742968eebbd2a41