Sort by relevance for CirrusSearch geosearch
When the parameter sort=relevance is added to the query, the result
will not be sorted by distance to the center but by relevance defined
by the CirrusSearch scoring.

The scores for the relevance sort a applied like done in the
CirrusSearch SearchRequestBuilder. It would need a bigger refactoring
if we would want to make better use of that Request-/QueryBuilder
instead of this copy.

Bug: T307208
Change-Id: If9c9b7096bf74d9591ef36b7629b3fea5865b037