Avoid count() when the actual count is not needed
This is not about performance. count() is cheap in modern PHP. This
is about readability. This code doesn't care about the actual number
of elements in these arrays, but only if the array is empty or not.

We discussed this in similar patches in other codebases and found that
a strict comparison with [] is most readable for most people.

Change-Id: I002dc5e5e26c71ca1d164841601eed3917ed5dae