Use actual `const` for class constant in Coord class
This is now possible in the PHP versions we use.

I also remove a series of trivial getters that aren't necessary any
more when we can use an actual constant, as well as the extra `static`
caching in these getters. The main reasoning is that these operations
are all cheap, because the array is a constant and known to be tiny.

Change-Id: I88c19d5e1d2ccc514bb476f12c3d21576ad6434e