Ensure coordinates are stored as floats
Everything is documented as floats, but when integers are provided,
either by a file or parser function (ex: {{coord|1|-1}}) as integers
they are passed through as integers. Cast to floats to have consistent
expectations about what types to expect.

Likely simply casting when creating the Coord object is sufficient,
but it seemed sane to cast to floats at the points where the
coordinates are initially parsed out as well.

Bug: T322327
Change-Id: Ibaf6048e6f6363c037c07555a48632b3510d479c