Drop effectively unused $wgDefaultGlobe configuration
The intention of this patch is not to remove support for other
globes. What I what to propose here is to effectively hard-code the
default to be "earth". This makes the code much more predictable.

The feature is not used on Wikimedia wikis, according to CodeSearch.
The only reason to ever change it is on e.g. a Mars wiki where users
want the {{#coordinates}} parser function to default to "mars" when
they don't manually specify a globe. Unfortunatelly the code starts
to behave very unpredictable when you do that. A lot of places
already hard-code "earth" without taking the $wgDefaultGlobe setting
into account. Note how especially the QueryGeoSearch API module
doesn't even allow searching for globes other than the default.

It's much, much better to always provide the globe when it's not
Earth. I believe it's fair to assume that all wikis already do that.
If not it's easy to fix. The {{#coordinates}} parser function is
usually wrapped in a template anyway.

Change-Id: I47a8e1ca041fcc8e9e2ec93f5bf9d78c6500599c