Fix missing/inconsistent default globe argument
The first two places are no-ops. We are just re-using the same value
as before via the Globe object created a few lines above.

But the hook code was really missing the globe. The validation was
done with a hard-coded "earth" but the Coord object created with the
$wgDefaultGlobe configuration. What if this changes? Luckily it never
did so far. It's always "earth" as well.

Change-Id: If5eb809ddde8df50349eb68bdad72f70886b5d64