Make the Coord class default to Earth
The idea is to pull the $wgDefaultGlobe global out of the Coord
constructor and make it a responsibility of the code that calls the
constructor. This doesn't change anything in production.

Turns out the possibility to call the Coord constructor without a
globe was almost entirely unused anyway. At least in production code.
The gt_globe database field isn't even nullable. Only tests tried to
make Coord objects without providing a globe. This is easy to fix in
the test setup.

Change-Id: Ib78948bd94b35dd90dd24fed91d4b37d0316488a