Fix incomplete globe support in QueryGeoSearchElastic API
This probably doesn't make any difference in production. Support for
different globes is heavily incomplete and defaults to Earth anyway
pretty much everywhere. But in case the globe is recorded in the
Elastica backend (which I believe it is) we should use it here and
not assume some default that's possibly different.

Unfortunately this code isn't covered by unit tests, as far as I can
tell.

Change-Id: Iec6a0733d315a822e39f3f400f6de26e4c030c63