Fix distance/distanceTo to work with globes other than Earth
I decided to go with the approach previously drafted in I2fc8988 and
not start using the Coord or Globe classes in the Math class, but keep
Math a pure, trivial helper with no dependencies.

Note it's not forbidden to use coordinates with different globes. The
first globe will be used and the other ignored. This is intentional.
It's not trivial to predict what happens when we start returning null
or throwing exceptions. I would like to do this in later patches.

Also note how the relevant usages already have checks in place to make
sure the globe is the same for both coordinates.

Bug: T160141
Change-Id: I6efee58a6f4b9497ea9cebf2e2d7f128a46e5e66