Calculate and document sensible default for maximum longitude
This is a direct follow up for what I started in Ib704441.

For unknown globes we allow values from -360 to +360. This is
intentional as the specific min and max are unknown and could be
anything from [-360, 0] to [0, +360].

For known globes we default to [0, +360] or [min, min+360]. This is
for convenience when providing additional $wgGlobes.

Also:
* Document the default for the East sign.
* Make the default sign more obvious by using +1 instead of 1.

Change-Id: Ia5e81fb8747742eb40f24fc496410fac04005839