Various code cleanups related to usage error tracking
Various code cleanups in tests and production code. This is mostly
to make the code more predictable and easier to read. No behavior
changes.

Some details:

* The value "track" never existed. The feature was added in 2012 via
commit f9b3bda7, see https://phabricator.wikimedia.org/rEGDAf9b3bda7.
The code requires "warn" for the tracking category, not "track". It
appears like this was a mistake in the patch.

* For the same reason we can use null as the default, instead of the
string "none". It means the same.

* Also make the elements in the $wgGeoDataWarningLevel array optional
while we are here.

* One test really only covers the "invalid region" case, nothing else.
Let's make this visible.

* Some assertions had a misplaced 3rd argument. This is a message
string, not a precision.

Depends-On: I9ce8740dc5452d2502442e5fe733ff7ded56d804
Change-Id: I270a8e8de30d0b6404112e876cc73c1b6d74f68d