Use native PHPUnit feature to name test cases
A @dataProvider in PHPUnit can have names for the individual test
cases. This is a feature that comes with PHPUnit. While we can do it
manually there is not much of an advantage to do so, I believe.

Also switch to the yield syntax in a few cases to further reduce the
likelyhood of mistakes happening. The yield syntax makes it actually
easier to find e.g. duplicate array keys.

Change-Id: I6551c921e25a937ee3649c2ff610a3c3c40764a0