Drop 12 year old load-test script
This was added in 2012 as part of a patch that added support for a
Sphinx backend, see I15475fd.

That was removed just a year later via I85f6110. The script was left
behind. While it might be possible to use it without Sphinx I don't
think that's very useful. The script is very old and can't do much
anyway. I believe we have better options to benchmark and profile
code now.

Change-Id: I62cef033baa3875161c28c990c2d7073d99271eb