Accept Globes as objects in many more places
Unfortunately the Coord class itself cannot be changed to hold a
Globe object, but must support the globe as a public string property.
But a lot of other places can easily accept a Globe object, at least
as an option. I believe this makes some of the code more readable
because we don't have to go back and forth between strings and
Globe objects so often.

This also adds a convenient Coord::sameGlobe function that can be
used in surprisingly many places.

What do you think?

Bug: T160141
Change-Id: Ica28f31fe85a7ce4110e0ea8bd18f8b24ae1c431